requirements.txt
mysql-connector-python==9.0.0
python-dotenv==1.0.1
colorama==0.4.6
